feat(tests): add unit tests for component code generation and graph code generation
