+++ src/backend/tests/unit/graph/graph/test_base.py
@@ -3,8 +3,6 @@ from collections import deque
 import pytest
 
 from langflow import components
-from langflow.code_gen.component import generate_instantiation_string, generate_script
-from langflow.code_gen.generic import generate_import_statement
 from langflow.graph.graph.base import Graph
 from langflow.graph.graph.constants import Finish
 
@@ -134,54 +132,3 @@ def test_graph_functional_start_end():
     assert len(results) == len(ids) + 1
     assert all(result.vertex.id in ids for result in results if hasattr(result, "vertex"))
     assert results[-1] == Finish()
-
-
-def test_generate_import_statement_and_instantiation_string():
-    chat_input_instance = components.inputs.ChatInput(_id="chatInput-1230")
-    import_statement = generate_import_statement(chat_input_instance)
-    instantiation_string = generate_instantiation_string(chat_input_instance)
-    assert import_statement == "from langflow.components.inputs import ChatInput"
-    assert instantiation_string == "chatinput_1230 = ChatInput(_id='chatInput-1230')"
-
-
-def test_generate_script():
-    chat_input = components.inputs.ChatInput(_id="chatInput-1230")
-    text_output = components.outputs.TextOutput.TextOutputComponent(_id="textoutput-1231")(
-        input_value=chat_input.message_response
-    )
-    script = generate_script(chat_input, text_output)
-    assert (
-        script
-        == """from langflow.components.inputs.ChatInput import ChatInput
-from langflow.components.outputs.TextOutput import TextOutputComponent
-
-chatinput_1230 = ChatInput(_id='chatInput-1230')
-textoutput_1231 = TextOutputComponent(_id='textoutput-1231')
-
-textoutput_1231(input_value=chatinput_1230.message_response)"""
-    )
-
-
-def test_gerenate_script_from_graph():
-    chat_input = components.inputs.ChatInput(_id="chatInput-1230")
-    text_output = components.outputs.TextOutput.TextOutputComponent(_id="textoutput-1231")(
-        input_value=chat_input.message_response
-    )
-    chat_output = components.outputs.ChatOutput(input_value="test", _id="chatOutput-1232")(
-        input_value=text_output.text_response
-    )
-    graph = Graph(chat_input, chat_output)
-    script = generate_script(*graph.sort_components())
-    assert (
-        script
-        == """from langflow.components.inputs.ChatInput import ChatInput
-from langflow.components.outputs.ChatOutput import ChatOutput
-from langflow.components.outputs.TextOutput import TextOutputComponent
-
-chatinput_1230 = ChatInput(_id='chatInput-1230')
-textoutput_1231 = TextOutputComponent(_id='textoutput-1231')
-chatoutput_1232 = ChatOutput(_id='chatOutput-1232')
-
-textoutput_1231(input_value=chatinput_1230.message_response)
-chatoutput_1232(input_value=textoutput_1231.text_response)"""
-    )
